requirements.txt
openai
streamlit
sounddevice
numpy
scipy
gtts
python-dotenv